+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade to `@commitspark/git-adapter-github` 0.7.0
 - Replace mandatory automatic entry ID generation with manual custom ID input when creating new entries
 - Refactor pages and layout to no longer cause branch list reload on every navigation
+- Move Git provider authentication to server and only expose encrypted tokens to client
 
 ## [0.11.0] - 2024-08-17